requirements.txt
opencv-python
numpy
Pillow
reportlab
PyPDF2
fitz
img2pdf





